AtCoder/ABC/005/C/ABC005C.go
package main

import (
	"bufio"
	"fmt"
	"os"
	"strconv"
)

// func main() {
func mainABC05C() {
	sc.Split(bufio.ScanWords)
	A, B := []int{}, []int{}

	T, N := nextInt(), nextInt()
	for i := 0; i < N; i++ {
		A = append(A, nextInt())
	}

	M := nextInt()
	for i := 0; i < M; i++ {
		B = append(B, nextInt())
	}

	fmt.Printf("%s\n", ABC05C(T, A, B))
}

// ABC05C ...
func ABC05C(T int, A, B []int) string {
	ai, bi, done := 0, 0, []int{}

	for ai < len(A) && bi < len(B) {
		if A[ai] > B[bi] {
			// no takoyaki
			bi++
		} else if B[bi]-A[ai] > T {
			// old takoyaki
			ai++
		} else {
			if A[ai]-B[bi] <= T || B[bi]-A[ai] <= T {
				done = append(done, B[bi])
			}
			ai++
			bi++
		}
	}

	if len(done) == len(B) {
		return "yes"
	}

	return "no"
}

var sc = bufio.NewScanner((os.Stdin))

func nextLine() string {
	sc.Scan()
	return sc.Text()
}

func nextInt() int {
	l := nextLine()
	i, e := strconv.Atoi(l)
	if e != nil {
		panic(e)
	}
	return i
}
